requirements.txt
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-openai>=0.0.5
langchain-text-splitters==0.3.8
langchain>=0.1.0
langsmith==0.3.32
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
numpy==1.26.4
openai==1.75.0
orjson==3.10.16
packaging==23.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.1
protobuf==4.25.6
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2023.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.32.0
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.34.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
